requirements.txt
pandas
numpy
scikit-learn
tensorflow===2.15.0
nltk
emoji
networkx
matplotlib
streamlit
keras===2.15.0
seaborn
speechrecognition
pytesseract
pillow